updates for login thru facebookand google
