+++ config/db.config.js
@@ -1,9 +1,7 @@
-const mysql=require('mysql2')
-const pool=mysql.createPool({
-host:'127.0.0.1',
-user:'root',
-database:'test',
-password:''
-})
-
-module.exports = pool.promise()
+module.exports={
+  database: 'test',
+  host: '127.0.0.1',
+  user: 'root',
+  password: '',
+  
+};
